backend/requirements.txt
Flask==2.0.3
pytesseract==0.3.8
Pillow==8.4.0
spacy==3.2.4
transformers==4.20.1
pymongo==3.12.1